# ui/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DeletedCounts": {
        "description": "Rows deleted by a project removal or a reset.",
        "properties": {
          "artifacts": {
            "format": "int64",
            "type": "integer"
          },
          "backfills": {
            "format": "int64",
            "type": "integer"
          },
          "events": {
            "format": "int64",
            "type": "integer"
          },
          "flows": {
            "format": "int64",
            "type": "integer"
          },
          "logs": {
            "format": "int64",
            "type": "integer"
          },
          "rules": {
            "format": "int64",
            "type": "integer"
          },
          "runs": {
            "format": "int64",
            "type": "integer"
          },
          "schedules": {
            "format": "int64",
            "type": "integer"
          },
          "task_runs": {
            "format": "int64",
            "type": "integer"
          },
          "variables": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "flows",
          "runs",
          "task_runs",
          "logs",
          "events",
          "artifacts",
          "schedules",
          "backfills",
          "rules",
          "variables"
        ],
        "type": "object"
      },
...
      "Id": {
        "type": "string"
      },
...
      "ResetBody": {
        "properties": {
          "backup": {
            "description": "Write a copy to `<home>/backups/` first; defaults to true.",
            "type": [
              "boolean",
              "null"
            ]
          },
          "scope": {
            "description": "`history` or `everything`; required.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "ResetResult": {
        "properties": {
          "backup_path": {
            "description": "The copy written before the reset, when one was asked for.",
            "type": [
              "string",
              "null"
            ]
          },
          "deleted": {
            "$ref": "#/components/schemas/DeletedCounts"
          },
          "scope": {
            "type": "string"
          }
        },
        "required": [
          "scope",
          "deleted"
        ],
        "type": "object"
      },
...
      "Run": {
        "properties": {
          "attempt": {
            "format": "int64",
            "type": "integer"
          },
          "backfill_id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "crash_count": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "created_at": {
            "$ref": "#/components/schemas/i64"
          },
          "created_by": {
            "type": "string"
          },
          "end_time": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/i64"
              }
            ]
          },
          "engine_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "engine_pid": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "external_id": {
            "$ref": "#/components/schemas/Id"
          },
          "failure_count": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "flow_id": {
            "format": "int64",
            "type": "integer"
          },
          "flow_name": {
            "type": "string"
          },
          "group": {
            "description": "The flow's group, read through the flow: never stored on the run.",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "parameters": {
            "type": "object"
          },
          "parent_run_id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "priority": {
            "format": "int64",
            "type": "integer"
          },
          "project": {
            "type": "string"
          },
          "report_seq": {
            "format": "int64",
            "type": "integer"
          },
          "schedule_id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "scheduled_time": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/i64"
              }
            ]
          },
          "start_time": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/i64"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/State"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "task_counts": {
            "additionalProperties": {
              "format": "int64",
              "type": "integer"
            },
            "description": "Task runs of this run counted by state type; empty until tasks exist.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "total_run_time": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/i64"
              }
            ]
          }
        },
        "required": [
          "id",
          "external_id",
          "flow_id",
          "flow_name",
          "project",
          "name",
          "state",
          "failure_count",
          "crash_count",
          "created_at"
        ],
        "type": "object"
      },
...
      "State": {
        "properties": {
          "details": {
            "type": "object"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "timestamp": {
            "$ref": "#/components/schemas/i64"
          },
          "type": {
            "$ref": "#/components/schemas/StateType"
          }
        },
        "required": [
          "type",
          "name",
          "timestamp"
        ],
        "type": "object"
      },
...
      "StateType": {
        "enum": [
          "Scheduled",
          "Pending",
          "Running",
          "Completed",
          "Failed",
          "Cancelled",
          "Crashed",
          "Paused",
          "Cancelling"
        ],
        "type": "string"
      },
...
      "TransitionBody": {
        "properties": {
          "details": {
            "type": "object"
          },
          "force": {
            "type": "boolean"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "$ref": "#/components/schemas/StateType"
          }
        },
        "required": [
          "type"
        ],
        "type": "object"
      },
      "TransitionRejected": {
        "properties": {
          "current": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/State"
              }
            ]
          },
          "error": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "reason"
        ],
        "type": "object"
      },
...
      "i64": {
        "format": "int64",
        "type": "integer"
      }
    }
  },
  "info": {
    "description": "Local-first pipeline orchestrator API",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "cereyan",
    "version": "1.14.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/database/reset": {
      "post": {
        "operationId": "reset_database",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetResult"
                }
              }
            },
            "description": ""
          },
          "409": {
            "description": "A reset is already running"
          },
          "422": {
            "description": "Missing or unknown scope"
          },
          "500": {
            "description": "The copy failed; nothing was deleted"
          }
        },
        "tags": [
          "database"
        ]
      }
    },
...
    "/api/runs/{id}": {
      "delete": {
        "operationId": "delete_run",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "get_run",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{id}/transition": {
      "post": {
        "operationId": "transition_run",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionRejected"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}